services/ingestion/requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
google-cloud-storage==2.18.2
google-cloud-logging==3.11.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
